Release: Breaking changes
This new release updates boiler to use new version of shiftschema (0.2.0) that introduces some breaking changes to validators and filters where previously validation context was being used, now we always get in a model and context is reserved for manually injecting additional validation context.

Because of this, we had to update our validators in user domain. You will need to update your filters and validators code as well when moving to this version.